// app/models/nhanvien/update-nhanvien-request.tsx
export interface UpdateNhanVienRequest {
  maNhanVien: string;
  hoTen: string;
  chucVu: string;
  mail: string;
  ngaySinh: string;
  soCCCD: string;
  ngayCap: string;
  queQuan: string;
  noiOHienTai: string;
  nguoiThanLienHe: string;
  soDienThoaiNguoiLienHe: string;
  stkNganHang: string;
  nganHang: string;
  maPhongBan: number;
  soDienThoai: string;
  idVanTay: number;
}